Refactor filters operation support users can build their own filter easily.
.
