+++ RecDP/pyrecdp/primitives/llmutils/filter.py
@@ -12,7 +12,6 @@ def filter_by_blocklist_spark(spark_df):
 
 
 def filter_by_blocklist(data_dir, out_dir, data_file_type="jsonl"):
-    from pyrecdp.LLM import ResumableTextPipeline
     from pyrecdp.primitives.operations import URLFilter
     from pyrecdp.LLM import ResumableTextPipeline
 
@@ -41,7 +40,6 @@ def filter_by_bad_words_spark(spark_df, language="en"):
 
 
 def filter_by_bad_words(data_dir, out_dir, data_file_type="jsonl", language="en"):
-    from pyrecdp.LLM import ResumableTextPipeline
     from pyrecdp.primitives.operations import BadwordsFilter
     from pyrecdp.LLM import ResumableTextPipeline
 
@@ -70,7 +68,6 @@ def filter_by_length_spark(spark_df, minimum_length=100, maximum_length=-1):
 
 
 def filter_by_length(data_dir, out_dir, data_file_type="jsonl", minimum_length=100, maximum_length=-1):
-    from pyrecdp.LLM import ResumableTextPipeline
     from pyrecdp.primitives.operations import LengthFilter
     from pyrecdp.LLM import ResumableTextPipeline
